Problem class added
The Problem class closely follows the formulation from AIMA. It contains the
logic which defines the state space.

Specifically given a state, it determines what actions are available from that
state, as well as what the result of taking that action will be.

The Problem class also contains the logic for checking if a State is a goal
state.

Lastly, it can create an initial State from an input formatted like the one on
the course webpage.